[#132] Remove command toggling attributes
I don't belive these options should exist, because editor commands need
to be explicitly typed in the prompt to be accessed. Because of that they don't
interfere with other commands or clutter the prompt.

If the amount of commands does clutter the command prompt, maybe there
should be less commands