Remove outputchannel from the constructor
Refactored the crashReporter module to not use outputchannel,
instead delegated the action to the gutters module. This action
reduces the parameters needed by the constructor function of the
crashReporter module. The capture function in the gutters module
is now being called by creating a new instance of the class.

There are no side effects associated with these changes.