Add `tslib` as runtime dep
https://github.com/ryanluker/vscode-coverage-gutters/issues/370#issuecomment-1120402680

According to [typescriptlang.org](https://www.typescriptlang.org/tsconfig#importHelpers),
not using `importHelpers` "can result in code duplication if the same helper is used in many different modules".
So use `tslib` might be a better solution.

The reason the unit tests are not failing is that
unit tests use the development environment's `node_modules` when running.

Tested on my computer.

close #370